Test : 플레이어 입력값 유효성검사 테스트 케이스 작성
1.페키지경로에 맞게 경로 수정
2.error message 테스트 추가
